Mostly finished layout, theming, and polish.
Still need to make orientation changes trigger a re-render for the console, in case it would be hidden by the height of the editor. Could also do this without an orientation change event, eg by just listening to a resize callback.

Also considering adding a tab for the simulator, or re-thinking how the console is laid out.

Can also turn the autoResize in ivygate off using this rerender method, but benefit is unclear.